Added Testcase - 4

+++ Statistics.Test/StatsUnitTest.cs
@@ -39,28 +39,25 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<double> {(double) double.NaN, (double) 1.5, (double) 8.9 });
+                new List<double> {double.NaN, (double) 1.5, (double) 8.9 });
             double epsilon = 0.001F;
             Assert.True(Math.Abs(computedStats.average - 5.2) <= epsilon);
             Assert.True(Math.Abs(computedStats.max - 8.9) <= epsilon);
             Assert.True(Math.Abs(computedStats.min - 1.5) <= epsilon);
         }
-        /*
+        
         [Fact]
         public void ReportsNaNForAllNaNInput()
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<float> {(float) Double.NaN, (float)Double.NaN, (float)Double.NaN });
-            // All fields of computedStats (average, max, min) must be
-            // Double.NaN (not-a-number), as described in
-            // https://docs.microsoft.com/en-us/dotnet/api/system.double.nan?view=netcore-3.1
-            // Specify the Assert statements here
-            Assert.True(computedStats.average.Equals(Double.NaN));
-            Assert.True(computedStats.max.Equals(Double.NaN));
-            Assert.True(computedStats.min.Equals(Double.NaN));
+                new List<double> {double.NaN, double.NaN, double.NaN });
+            Assert.True(computedStats.average.Equals(double.NaN));
+            Assert.True(computedStats.max.Equals(double.NaN));
+            Assert.True(computedStats.min.Equals(double.NaN));
         }
 
+        /*
         [Fact]
         public void ReportsNaNForAbsurdInput()
         {
